requirements.txt
streamlit
tensorflow
matplotlib
streamlit-option-menu
pillow
streamlit-drawable-canvas
pickle5